+++ martypy/ClientMV2.py
@@ -61,6 +61,8 @@ class ClientMV2(ClientGeneric):
         self.ricHwElemsInfoByIDNo = {}
         self.ricHwElemsList = []
         self.loggingCallback = None
+        self.publishedMsgCallback = None
+        self.reportMsgCallback = None
         self._initComplete = False
         self._minSysVersForSubscribeAPI = "1.0.0"
         self._interfaceMethod = method
@@ -119,20 +121,14 @@ class ClientMV2(ClientGeneric):
             return
         self.isClosing = True
         # Stop any publishing messages
-        if self._systemVersionGtEq(self._minSysVersForSubscribeAPI):
-            self.ricIF.sendRICRESTCmdFrame('{"cmdName":"subscription","action":"update",' + \
-                '"pubRecs":[' + \
-                    '{"name":"MultiStatus","rateHz":0},' + \
-                    '{"name":"PowerStatus","rateHz":0},' + \
-                    '{"name":"AddOnStatus","rateHz":0}' + \
-                ']}')
-            # Allow message to be sent
-            time.sleep(0.5)
+        self._unsubscribeFromPubMessages()
         # Close the RIC interface
         self.ricIF.close()
 
     def wait_if_required(self, expected_wait_ms: int, blocking_override: Union[bool, None]):
         if not self.is_blocking(blocking_override):
+            return
+        if self.subscribeRateHz <= 0:
             return
 
         deadline = time.time() + expected_wait_ms/1000 + self.max_blocking_wait_time
@@ -512,7 +508,52 @@ class ClientMV2(ClientGeneric):
         return result
 
     def register_logging_callback(self, loggingCallback: Callable[[str],None]) -> None:
+        '''
+        Register a callback function to be called on every log message from RIC.
+        Log messages are used mainly for debugging and error reporting.
+        Args:
+            messageCallback: a callback function (with one string argument - the log message)
+        Returns:
+            None
+        '''
         self.loggingCallback = loggingCallback
+
+    def register_publish_callback(self, messageCallback: Callable[[int],None]) -> None:
+        '''
+        Register a callback function to be called on every message published by RIC.
+        RIC publishes information like the accelerometer and joint positions constantly.
+        If registered, the callback is called after the message is fully decoded, so a common
+        use-case is to check the topic (the int passed to the callback) to see if the information
+        is of interest (for example topic == Marty.PUBLISH_TOPIC_ACCELEROMETER if new accelerometer
+        data is available). Then get the changed information using the regular get_accelerometer method
+        (or get_joints, etc for other data).
+        Args:
+            messageCallback: a callback function (with one argument - the topic code of the published
+                    message) that will be called on every message published by RIC.
+        Returns:
+            None
+        '''
+        self.publishedMsgCallback = messageCallback
+
+    def register_report_callback(self, messageCallback: Callable[[str],None]) -> None:
+        '''
+        Register a callback function to be called on every report message recieved from RIC.
+        Report messages are used for alert conditions such as Falling, Over Current, etc.
+        The report message (passed in the callback function str) is a JSON string.
+        The elements of the JSON string include:
+            msgType: the type of report message (this will be "warn" for all alerts)
+            msgBody: the message text (this is "freeFallDet" or "overCurrentDet")
+            IDNo: the IDNo of the element that generated the report (the codes for these
+                are in the Marty.HW_ELEM_IDS dictionary but additional ones may appear if
+                add-ons create warnings - the IDNos of add-ons are not fixed numbers but you
+                can see the IDNo of add-ons by using the get_add_ons_status() method)
+        Args:
+            messageCallback: a callback function (with one string argument)
+                    that will be called on every message published by RIC.
+        Returns:
+            None
+        '''
+        self.reportMsgCallback = messageCallback
 
     def get_interface_stats(self) -> Dict:
         ricIFStats = self.ricIF.getStats()
@@ -530,17 +571,35 @@ class ClientMV2(ClientGeneric):
             # logger.debug(f"ROSSERIAL message received {len(decodedMsg.payload)}")
             self.lastRICSerialMsgTime = time.time()
             if decodedMsg.payload:
-                RICROSSerial.decode(decodedMsg.payload, 0, self.ricHardware.updateWithROSSerialMsg)
+                RICROSSerial.decode(decodedMsg.payload, 0, self._rxPublishedMsg)
         elif decodedMsg.protocolID == RICProtocols.PROTOCOL_RICREST:
             # logger.debug(f"RIC REST message received {decodedMsg.payload}")
-            pass
+            # Callback for REPORT messages
+            if (self.reportMsgCallback is not None) and (decodedMsg.msgTypeCode == RICProtocols.MSG_TYPE_REPORT):
+                try:
+                    self.reportMsgCallback(decodedMsg.payload)
+                except:
+                    logger.exception("Report callback failed:")
         else:
-            # logger.debug(f"RIC OTHER message received {decodedMsg.payload}")
+            logger.debug(f"RIC UNKNOWN message received {decodedMsg.payload}")
             pass
+
+    def _rxPublishedMsg(self, topicID: int, payload: bytes):
+        # Decode the message
+        self.ricHardware.updateWithROSSerialMsg(topicID, payload)
+        # Callback on published messages
+        if self.publishedMsgCallback is not None:
+            try:
+                self.publishedMsgCallback(topicID)
+            except:
+                logger.exception("Publish callback failed:")
 
     def _logDebugMsg(self, logMsg: str) -> None:
         if self.loggingCallback:
-            self.loggingCallback(logMsg)
+            try:
+                self.loggingCallback(logMsg)
+            except:
+                logger.exception("Logging callback failed:")
 
     def _msgTimerCallback(self) -> None:
         if self.isClosing:
@@ -589,6 +648,18 @@ class ClientMV2(ClientGeneric):
                                 '{' + f'"name":"AddOnStatus","rateHz":{self.subscribeRateHz}' + '}' + \
                             ']}')
             self.lastSubscrReqMsgTime = time.time()
+
+    def _unsubscribeFromPubMessages(self):
+        if self._systemVersionGtEq(self._minSysVersForSubscribeAPI):
+            # Send unsubscribe request
+            self.ricIF.sendRICRESTCmdFrame('{"cmdName":"subscription","action":"update",' + \
+                '"pubRecs":[' + \
+                    '{"name":"MultiStatus","rateHz":0},' + \
+                    '{"name":"PowerStatus","rateHz":0},' + \
+                    '{"name":"AddOnStatus","rateHz":0}' + \
+                ']}')
+            # Allow message to be sent
+            time.sleep(0.5)
 
     def _onReconnect(self):
         self._subscribeToPubMessages(True)